Move Level 2 BUFR and netCDF readers into level2 module.
